requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
kaleido==0.2.1
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.47
langchain-experimental==0.3.4
langchain-google-genai==2.1.1
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
matplotlib==3.10.1
matplotlib-inline==0.1.7
pymongo==4.11.3
python-dotenv==1.0.1
requests==2.32.3
seaborn==0.13.2
selenium==4.30.0
sentence-transformers==4.0.1
vnstock==3.2.2
